BACKLOG-11925: split picker dialog into leftPanel and MainPanel
