+++ src/javascript/EditPanel/EditPanelContent/FormBuilder/Section/FieldSet/Field/SelectorTypes/Picker/Picker.utils.spec.js
@@ -1,4 +1,4 @@
-import {allSitesEntry, getSiteNodes, extractConfigs} from './Picker.utils';
+import {allSitesEntry, getSiteNodes, extractConfigs, getSite, getDetailedPathArray, getPathWithoutFile} from './Picker.utils';
 
 describe('Picker utils', () => {
     describe('getSiteNodes', () => {
@@ -47,6 +47,10 @@ describe('Picker utils', () => {
             expect(getSiteNodes(data([siteB, siteA]), allSitesLabel)).toStrictEqual([allSitesEntry(allSitesLabel), siteA, siteB]);
         });
 
+        it('should return empty array when data jcr doesn\'t exist', () => {
+            expect(getSiteNodes({}, allSitesLabel)).toStrictEqual([]);
+        });
+
         it('should add all sites entry when more than 2 sites', () => {
             expect(getSiteNodes(data([siteA, siteB]), allSitesLabel)).toStrictEqual([allSitesEntry(allSitesLabel), siteA, siteB]);
             expect(getSiteNodes(data([siteA]), allSitesLabel)).toStrictEqual([siteA]);
@@ -91,6 +95,39 @@ describe('Picker utils', () => {
                 },
                 type: 'files'
             }]);
+        });
+    });
+
+    describe('getPathWithoutFile', () => {
+        it('should return undefined when initialSelectedItem is empty', () => {
+            expect(getPathWithoutFile()).toBe(undefined);
+        });
+
+        it('should return /toto/tata when give a cat.js file', () => {
+            expect(getPathWithoutFile('/toto/tata/cat.js')).toBe('/toto/tata');
+        });
+    });
+
+    describe('getSite', () => {
+        it('should return undefined if path is not defined', () => {
+            expect(getSite()).toBe(undefined);
+        });
+
+        it('should return /site/digitall when give a full path', () => {
+            expect(getSite('/site/digitall/files/cats/cats.js')).toBe('/site/digitall');
+        });
+    });
+
+    describe('getDetailedPathArray', () => {
+        it('should return [] if path is not defined', () => {
+            expect(getDetailedPathArray()).toEqual([]);
+        });
+
+        it('should return detailed path if path is not defined', () => {
+            expect(getDetailedPathArray('/site/digitall/files/cats/cats.js', '/site/digitall')).toEqual([
+                '/site/digitall',
+                '/site/digitall/files'
+            ]);
         });
     });
 });
