+++ src/javascript/EditPanel/EditPanelContent/FormBuilder/Section/FieldSet/Field/SelectorTypes/Picker/PickerDialog/LeftPanel.jsx
@@ -1,0 +1,115 @@
+import React, {useState} from 'react';
+import PropTypes from 'prop-types';
+
+import Drawer from '@material-ui/core/Drawer';
+import {NodeTrees, PickerTreeViewMaterial} from '@jahia/react-material';
+import SiteSwitcher from '~/DesignSystem/SiteSwitcher/SiteSwitcher';
+import {Picker} from '@jahia/react-apollo';
+
+import {withStyles} from '@material-ui/core';
+
+import {getDetailedPathArray, getSite} from '../Picker.utils';
+
+const styles = theme => ({
+    drawerPaper: {
+        width: '15vw',
+        backgroundColor: theme.palette.ui.beta
+    },
+    listItem: {
+        color: theme.palette.primary.contrastText
+    }
+});
+
+const LeftPanelCmp = ({
+    site,
+    siteNodes,
+    field,
+    lang,
+    initialSelectedItem,
+    nodeTreeConfigs,
+    onSelectSite,
+    selectedPath,
+    setSelectedPath,
+    classes
+}) => {
+    const selectedItemSite = getSite(initialSelectedItem);
+    const initialPathOpenPath = getDetailedPathArray(initialSelectedItem, selectedItemSite);
+    const [openPaths, setOpenPaths] = useState(initialPathOpenPath);
+    const showSiteSwitcher = !(field.selectorOptions && field.selectorOptions.find(option => option.value === 'site'));
+
+    const openPath = path => {
+        setOpenPaths(previousOpenPaths => previousOpenPaths.concat([path]));
+    };
+
+    const closePath = path => {
+        setOpenPaths(previousOpenPaths => previousOpenPaths.filter(item => item !== path));
+    };
+
+    return (
+        <Drawer
+            open
+            component="nav"
+            variant="permanent"
+            anchor="left"
+            classes={{
+                paper: classes.drawerPaper
+            }}
+        >
+            {showSiteSwitcher &&
+            <SiteSwitcher
+                id="site-switcher"
+                siteKey={site}
+                siteNodes={siteNodes}
+                onSelectSite={siteNode => {
+                    const path = onSelectSite(siteNode);
+                    setOpenPaths([path]);
+                    setSelectedPath(path);
+                }}
+            />}
+
+            <NodeTrees path={selectedPath}
+                       rootPath="/"
+                       siteKey={site}
+                       nodeTreeConfigs={nodeTreeConfigs}
+            >
+                {({path, rootPath, openableTypes, selectableTypes, rootLabel, setRefetch, dataCmRole}) => (
+                    <Picker rootPaths={[rootPath]}
+                            openPaths={openPaths}
+                            openableTypes={openableTypes}
+                            selectableTypes={selectableTypes}
+                            queryVariables={{lang: lang}}
+                            selectedPaths={[path]}
+                            openSelection={false}
+                            setRefetch={setRefetch}
+                            onOpenItem={(path, open) => (open ? openPath(path) : closePath(path))}
+                            onSelectItem={setSelectedPath}
+                    >
+                        {({handleSelect, ...others}) => (
+                            <PickerTreeViewMaterial {...others}
+                                                    dataCmRole={dataCmRole}
+                                                    rootLabel={rootLabel}
+                                                    classes={{listItem: classes.listItem}}
+                        />
+                    )}
+                    </Picker>
+            )}
+            </NodeTrees>
+        </Drawer>
+    );
+};
+
+LeftPanelCmp.propTypes = {
+    site: PropTypes.string.isRequired,
+    siteNodes: PropTypes.array.isRequired,
+    field: PropTypes.object.isRequired,
+    lang: PropTypes.string.isRequired,
+    initialSelectedItem: PropTypes.array,
+    nodeTreeConfigs: PropTypes.array.isRequired,
+    selectedPath: PropTypes.string,
+    setSelectedPath: PropTypes.func.isRequired,
+    onSelectSite: PropTypes.func.isRequired,
+    classes: PropTypes.object.isRequired
+};
+
+export const LeftPanel = withStyles(styles)(LeftPanelCmp);
+LeftPanel.displayName = 'LeftPanel';
