--- src/javascript/EditPanel/EditPanelContent/FormBuilder/Section/FieldSet/Field/SelectorTypes/Picker/PickerDialog/PickerDialog/index.js
@@ -1,1 +1,0 @@
-export {PickerDialog} from './PickerDialog';
